feat: Added paging buttons to PostsMain

+++ NoFilterForum/Models/ViewModels/PostsViewModel.cs
@@ -8,8 +8,8 @@ namespace NoFilterForum.Models.ViewModels
         public List<PostDataModel> Posts { get; set; }
         public string Title { get; set; }
         public int Page { get; set; }
-        public int TotalPages { get; set; }
-        public PostsViewModel(UserDataModel currentUser, List<PostDataModel> posts,string title,int page, int totalPages) 
+        public double TotalPages { get; set; }
+        public PostsViewModel(UserDataModel currentUser, List<PostDataModel> posts,string title,int page, double totalPages) 
         { 
             CurrentUser = currentUser;
             Posts = posts;
